Refactor: move OAuth2 models into own folder & split into multiple files.
