+++ src/model/mod.rs
@@ -21,5 +21,3 @@ pub mod alliance;
 pub mod character;
 pub mod corporation;
 pub mod oauth2;
-
-mod util;
